[refactor] Hub application 레이어에서 service 패키지 제거
