+++ hub/src/main/java/sparta/rooibos/hub/application/port/in/HubService.java
@@ -1,4 +1,4 @@
-package sparta.rooibos.hub.application.service.port.in;
+package sparta.rooibos.hub.application.port.in;
 
 import sparta.rooibos.hub.application.dto.hub.request.CreateHubRequest;
 import sparta.rooibos.hub.application.dto.hub.request.SearchHubRequest;
@@ -15,7 +15,7 @@ public interface HubService {
 
     GetHubResponse getHub(UUID hubId);
 
-    UUID getHubIdByRegion(String region);
+    GetHubResponse getHubByRegion(String region);
 
     SearchHubResponse searchHub(SearchHubRequest searchHubRequest);
